backend/requirements.txt
django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
python-dotenv>=1.0
psycopg2-binary>=2.9
django-cors-headers>=4.0